# This is a combination of 2 commits. # The first commit's message is:
gsk: gsk_render_node_set_transform() => GskTransformNode

Instead of having a setter for the transform, have a GskTransformNode.

Most of the oprations that GTK does do not require a transform, so it
doesn't make sense to have it as a primary attribute.

Also, changing the transform requires updating the uniforms of the GL
renderer, so we're happy if we can avoid that.

# The 2nd commit message will be skipped:

#	xxx: 81ba6e6d1ebdd00b5562dd51749973e3296f05e6